Refactored Day 15 to new format
