--- .github/workflows/2019_day15.yml
@@ -1,29 +1,0 @@
-name: 2019 - Day 15
-
-on:
-  push:
-    branches: [ master ]
-    paths: "2019/Day15/**"
-  pull_request:
-    branches: [ master ]
-    paths: "2019/Day15/**"
-
-jobs:
-  Build_and_Test_2019_Day15:
-
-    runs-on: windows-latest
-
-    steps:
-    - uses: actions/checkout@v2
-    - name: Setup .NET Core
-      uses: actions/setup-dotnet@v1
-      with:
-        dotnet-version: 3.1.101
-    - name: Install dependencies for Solution
-      run: dotnet restore 2019/Day15/Solution/Day15.csproj
-    - name: Build
-      run: dotnet build --configuration Release --no-restore 2019/Day15/Solution/Day15.csproj
-    - name: Install dependencies for Tests
-      run: dotnet restore 2019/Day15/Tests/Day15Tests.csproj
-    - name: Test
-      run: dotnet test --no-restore --verbosity normal 2019/Day15/Tests/Day15Tests.csproj
