--- 2019/Day15/Solution/Day15.csproj
@@ -1,18 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-    <UserSecretsId>863a8a04-731a-4ccf-b330-6cb7444b8d0e</UserSecretsId>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="naochelper" Version="1.2.5" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\IntCodeMachine\Solution\IntCodeMachine.csproj" />
-  </ItemGroup>
-
-</Project>
