--- 2019/Day15/Solution/DroidInstructions.cs
@@ -1,10 +1,0 @@
-namespace Day15
-{
-    public enum Instruction
-    {
-        North = 1,
-        South = 2,
-        West = 3,
-        East = 4
-    }
-}
