--- 2019/Day15/Solution/DroidResponses.cs
@@ -1,9 +1,0 @@
-namespace Day15
-{
-    public enum Response
-    {
-        Wall = 0,
-        Moved = 1,
-        Oxygen = 2
-    }
-}
