--- 2019/Day15/Solution/appsettings.json
@@ -1,5 +1,0 @@
-{
-    "Secrets": {
-        "Cookie": "Configured in User Secrets"
-    }
-}
