+++ 2019/Solutions/Day15/Enums.cs
@@ -1,0 +1,17 @@
+﻿namespace Solutions.Day15
+{
+	public enum Instruction
+	{
+		North = 1,
+		South = 2,
+		West = 3,
+		East = 4
+	}
+	
+    public enum Response
+    {
+        Wall = 0,
+        Moved = 1,
+        Oxygen = 2
+    }
+}
