+++ 2019/Tests/Day15Tests/Day15Tests.cs
@@ -1,0 +1,6 @@
+﻿namespace Tests.Day15Tests
+{
+	public class Day15Tests
+	{
+	}
+}
